Everything almost working!
We have yet to finish off the finances tab in the admin view as well as fixing the invite system where the user has to close the app and open the app to enter the main flow of the app. We also have issues with the creator of the app changing the code of their own band. Manager dashboard view needs to be implemented and then we can look at finally releasing to the app store!